[chore][#84] 공통 사용 모듈 정리
